requirements.txt
gTTS==2.4.0
huggingface_hub==0.16.4
llama_cpp_python==0.1.78
loguru==0.7.2
numpy==1.24.4
openai==1.3.7
opencv_python==4.8.1.78
Pillow==10.1.0
streamlit==1.28.1
torch==2.1.0
transformers==4.35.0
